requirements.txt
django
djangorestframework
psycopg2
drf_yasg
django-cors-headers